finalizing of pep8 and pylint inside a config module we having ignore codes as "E0211,E0213,E1101,W0231,E0239,C0103,W0401"
